requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --no-hashes --output-file=requirements.txt
-e .
alembic==1.16.5
    # via flask-migrate
blinker==1.9.0
    # via flask
click==8.2.1
    # via flask
colorama==0.4.6 ; sys_platform == 'win32'
    # via
    #   click
    #   pytest
coverage==7.10.6
    # via pytest-cov
dnspython==2.7.0
    # via email-validator
email-validator==2.3.0
    # via journal
flask==3.1.2
    # via
    #   flask-login
    #   flask-migrate
    #   flask-sqlalchemy
    #   flask-wtf
    #   journal
flask-login==0.6.3
    # via journal
flask-migrate==4.1.0
    # via journal
flask-sqlalchemy==3.1.1
    # via
    #   flask-migrate
    #   journal
flask-wtf==1.2.2
    # via journal
greenlet==3.2.4 ; (python_full_version < '3.14' and platform_machine == 'AMD64') or (python_full_version < '3.14' and platform_machine == 'WIN32') or (python_full_version < '3.14' and platform_machine == 'aarch64') or (python_full_version < '3.14' and platform_machine == 'amd64') or (python_full_version < '3.14' and platform_machine == 'ppc64le') or (python_full_version < '3.14' and platform_machine == 'win32') or (python_full_version < '3.14' and platform_machine == 'x86_64')
    # via sqlalchemy
gunicorn==23.0.0
    # via journal
idna==3.10
    # via email-validator
iniconfig==2.1.0
    # via pytest
itsdangerous==2.2.0
    # via
    #   flask
    #   flask-wtf
jinja2==3.1.6
    # via flask
mako==1.3.10
    # via alembic
markdown==3.8.2
    # via journal
markupsafe==3.0.2
    # via
    #   flask
    #   jinja2
    #   mako
    #   werkzeug
    #   wtforms
mypy==1.17.1
mypy-extensions==1.1.0
    # via mypy
packaging==25.0
    # via
    #   gunicorn
    #   pytest
pathspec==0.12.1
    # via mypy
pluggy==1.6.0
    # via
    #   pytest
    #   pytest-cov
pygments==2.19.2
    # via pytest
pytest==8.4.1
    # via pytest-cov
pytest-cov==6.2.1
python-dotenv==1.1.1
    # via journal
ruff==0.12.11
sqlalchemy==2.0.43
    # via
    #   alembic
    #   flask-sqlalchemy
typing-extensions==4.15.0
    # via
    #   alembic
    #   mypy
    #   sqlalchemy
werkzeug==3.1.3
    # via
    #   flask
    #   flask-login
    #   journal
wtforms==3.2.1
    # via flask-wtf
